+++ CHANGES.md
@@ -3,4 +3,9 @@
 
 To be released.
 
-- Implement `leftpad.func.leftpad`.
+- Add CLI. [#3]
+- Implement `leftpad.func.leftpad`. [#1][#2]
+
+[#1]: https://github.com/admire93/pyleftpad/issues/1
+[#2]: https://github.com/admire93/pyleftpad/pull/2
+[#3]: https://github.com/admire93/pyleftpad/issues/3